Editing servers, startcmd
On the previous commits, the startcmd was put togheter when the server
was created. As someone might want to use a different port, then editing
the startup cmd might go wrong, so I'll keep the {port} tag in it, which
fetches the port data from SQL and applies it every time it starts the
server.