Improvements in server checking, name, tables.
Name changed to 1fx. # Server Panel.
Improvements in checking status - UDP could've timed out on mapswitches
or when the limit was full, which made the server restart without a
reason. So UDP query is removed, I added a lock checker, checks it twice
because it might be unlocked when server is changing maps, 5-sec timeout
should do the trick), if the lock exists, the server runs.
All tables made to be responsive (upon size change, they will just be
scrollable).